+++ utoipa-gen/CHANGELOG.md
@@ -50,6 +50,7 @@
 
 ### Changed
 
+* Chore enhance generic schema collection (https://github.com/juhaku/utoipa/pull/1116)
 * Enhance file uploads (https://github.com/juhaku/utoipa/pull/1113)
 * Move `schemas` into `ToSchema` for schemas (https://github.com/juhaku/utoipa/pull/1112)
 * Refactor `KnownFormat`